tests UPDATE merge ssh auth tests into one
